cleaning v3 code structure
