+++ src/v3/behaviours/index.ts
@@ -1,0 +1,3 @@
+export { when } from "./when";
+export { Behaviours } from "./behaviours";
+export type { Behaviour } from "./behaviours";
