+++ src/v3/hasher.ts
@@ -1,0 +1,3 @@
+import hash from "node-object-hash";
+
+export const hasher = hash({ coerce: { set: true, symbol: true } });
