+++ src/v3/mocks/index.ts
@@ -1,0 +1,4 @@
+export { mockAbstract } from "./mockAbstract";
+export { mockFunction } from "./mockFunction";
+export { mockType, mockInterface } from "./mockTypes";
+export { mockClass } from "./mockClass";
