+++ src/v3/spies/index.ts
@@ -1,0 +1,2 @@
+export { verifyThat } from "./verifyThat";
+export { spyMockedFunction } from "./mockedFunction";
